+++ src/i18n/missing/da.ts
@@ -1,10 +1,20 @@
 const da = {
 
+  accessory: {
+    available: '%s is available', // accessory name
+    statusActive: '%s is active', // accessory name
+    statusInactive: '%s is inactive', // accessory name
+    unavailable: '%s is not available', // accessory name
+  },
+
   config: {
 
     title: {
       minimum: 'Minimum',
       minimumStep: 'Step',
+      topicGetAvailable: 'Get Available/Reachable',
+      topicGetStatusActive: 'Get Active Status',
+      valueStatusActive: 'Status Active',
     },
   },